// admin-panel/src/api/admin/types.ts
export type AdminPhysicalFile = {
  id: number;
  slug: string;
  hashAlgo: string;
  fileHash: string;
  preHash: string;
  fileSize: number;
  mimeType: string;
  storagePath: string;
  status: string;
  createdAt: number;
  userFileCount: number;
  mediaItemCount: number;
};